Code refactoring + main simplified
